Disable docdb, enable rds instance
